Renamed classes & variables.
