+++ ts/Grid/Core/Table/ColumnResizing/AdjacentResizingMode.ts
