+++ ts/Grid/Core/Table/ColumnResizing/IndependentResizingMode.ts
