+++ ts/masters-grid/grid-lite.src.ts
@@ -21,7 +21,7 @@ import type _Options from '../Grid/Core/Options.ts';
 
 import AST from '../Core/Renderer/HTML/AST.js';
 import Templating from '../Core/Templating.js';
-import ColumnDistribution from '../Grid/Core/Table/ColumnDistribution/ColumnDistribution.js';
+import ColumnResizing from '../Grid/Core/Table/ColumnResizing/ColumnResizing.js';
 import DataConnector from '../Data/Connectors/DataConnector.js';
 import DataConverter from '../Data/Converters/DataConverter.js';
 import DataCursor from '../Data/DataCursor.js';
@@ -61,7 +61,7 @@ declare global {
         Grid: typeof _Grid;
         grid: typeof _Grid.grid;
         grids: Array<(_Grid|undefined)>;
-        ColumnDistribution: typeof ColumnDistribution;
+        ColumnResizing: typeof ColumnResizing;
         DataConverter: typeof DataConverter;
         DataCursor: typeof DataCursor;
         DataModifier: typeof DataModifier;
@@ -97,7 +97,7 @@ G.DataConverter = DataConverter;
 G.Grid = _Grid;
 G.grid = _Grid.grid;
 G.grids = _Grid.grids;
-G.ColumnDistribution = ColumnDistribution;
+G.ColumnResizing = ColumnResizing;
 G.DataModifier = DataModifier;
 G.DataPool = DataPool;
 G.DataTable = DataTable;
